refactor(sequencer): Make tests more readable
