+++ apps/sequencer/src/testing/sequencer_state.rs
@@ -14,10 +14,7 @@ use feed_registry::registry::{
     new_feeds_meta_data_reg_from_config, AllFeedsReports, FeedAggregateHistory,
 };
 use prometheus::metrics::FeedsMetrics;
-use tokio::sync::{
-    mpsc::{self, UnboundedReceiver},
-    RwLock,
-};
+use tokio::sync::{mpsc, RwLock};
 use utils::logging::init_shared_logging_handle;
 
 use crate::{
@@ -93,51 +90,4 @@ pub async fn create_sequencer_state_from_sequencer_config_file(
         feeds_management_cmd_send,
         blockchain_db: Arc::new(RwLock::new(InMemDb::new())),
     })
-}
-
-// Exclusively take config structure to init sequencer_state
-pub async fn create_sequencer_state_from_sequencer_config(
-    sequencer_config: SequencerConfig,
-    metrics_prefix: &str,
-) -> (
-    web::Data<SequencerState>,
-    UnboundedReceiver<(String, String)>,
-) {
-    let providers = init_shared_rpc_providers(&sequencer_config, Some(metrics_prefix)).await;
-
-    let log_handle = init_shared_logging_handle("INFO", false);
-
-    let (vote_send, vote_recv) = mpsc::unbounded_channel();
-    let (feeds_management_cmd_send, _feeds_management_cmd_recv) = mpsc::unbounded_channel();
-
-    let (_, feeds_config) = get_sequencer_and_feed_configs();
-
-    (
-        web::Data::new(SequencerState {
-            registry: Arc::new(RwLock::new(new_feeds_meta_data_reg_from_config(
-                &feeds_config,
-            ))),
-            reports: Arc::new(RwLock::new(AllFeedsReports::new())),
-            providers,
-            log_handle,
-            reporters: init_shared_reporters(&sequencer_config, Some(metrics_prefix)),
-            feed_id_allocator: Arc::new(RwLock::new(None)),
-            voting_send_channel: vote_send,
-            feeds_metrics: Arc::new(RwLock::new(
-                FeedsMetrics::new(metrics_prefix).expect("Failed to allocate feed_metrics"),
-            )),
-            active_feeds: Arc::new(RwLock::new(
-                feeds_config
-                    .feeds
-                    .into_iter()
-                    .map(|feed| (feed.id, feed))
-                    .collect(),
-            )),
-            sequencer_config: Arc::new(RwLock::new(sequencer_config.clone())),
-            feed_aggregate_history: Arc::new(RwLock::new(FeedAggregateHistory::new())),
-            feeds_management_cmd_send,
-            blockchain_db: Arc::new(RwLock::new(InMemDb::new())),
-        }),
-        vote_recv,
-    )
 }
